Adding support for Dashboard import/export via uid or folder

+++ api/dashboards.go
@@ -6,6 +6,7 @@ import (
 	"fmt"
 	"io/ioutil"
 	"os"
+	"path/filepath"
 	"regexp"
 	"strings"
 
@@ -26,12 +27,6 @@ type DashboardFilter struct {
 	DashFilter   string //name of dashboard
 }
 
-var quoteRegex *regexp.Regexp
-
-func init() {
-	quoteRegex, _ = regexp.Compile("['\"]+")
-}
-
 //GetFolders splits the comma delimited folder list and returns a slice
 func (s *DashboardFilter) GetFolders() []string {
 	if s.FolderFilter == "" {
@@ -40,6 +35,24 @@ func (s *DashboardFilter) GetFolders() []string {
 	s.FolderFilter = quoteRegex.ReplaceAllString(s.FolderFilter, "")
 
 	return strings.Split(s.FolderFilter, ",")
+}
+
+func (s DashboardFilter) ValidateDashboard(dashUid string) bool {
+	if s.DashFilter == "" {
+		return true
+	}
+	return dashUid == s.DashFilter
+}
+
+func (s DashboardFilter) Validate(folder, dashUid string) bool {
+	return s.ValidateDashboard(dashUid) && s.ValidateFolder(folder)
+}
+
+func (s DashboardFilter) ValidateFolder(folder string) bool {
+	if s.FolderFilter == "" {
+		return true
+	}
+	return folder == s.FolderFilter
 }
 
 //Update the slug in the board returned
@@ -52,13 +65,18 @@ func updateSlug(board *sdk.FoundBoard) {
 
 //ListDashboards: List all dashboards optionally filtered by folder name. If folderFilters
 // is blank, defaults to the configured Monitored folders
-func ListDashboards(client *sdk.Client, filters *DashboardFilter, query string) []sdk.FoundBoard {
+func ListDashboards(client *sdk.Client, filters *DashboardFilter) []sdk.FoundBoard {
 	ctx := context.Background()
 	var boardsList []sdk.FoundBoard = make([]sdk.FoundBoard, 0)
-	boardLinks, err := client.SearchDashboards(ctx, query, false)
+	boardLinks, err := client.SearchDashboards(ctx, "", false)
 	if err != nil {
 		panic(err)
 	}
+	//Fallback on defaults
+	if filters == nil {
+		filters = &DashboardFilter{}
+	}
+
 	folderFilters := filters.GetFolders()
 	var validFolder bool = false
 	var validUid bool = false
@@ -96,7 +114,7 @@ func ListDashboards(client *sdk.Client, filters *DashboardFilter, query string) 
 }
 
 //ImportDashboards saves all dashboards matching query to configured location
-func ImportDashboards(client *sdk.Client, query string, conf *viper.Viper) []string {
+func ImportDashboards(client *sdk.Client, filter DashboardFilter, conf *viper.Viper) []string {
 	var (
 		boardLinks []sdk.FoundBoard
 		rawBoard   []byte
@@ -105,11 +123,7 @@ func ImportDashboards(client *sdk.Client, query string, conf *viper.Viper) []str
 	)
 	ctx := context.Background()
 
-	filters := DashboardFilter{
-		DashFilter: strings.Join(config.GetDefaultGrafanaConfig().GetMonitoredFolders(), ","),
-	}
-
-	boardLinks = ListDashboards(client, &filters, query)
+	boardLinks = ListDashboards(client, &filter)
 	var boards []string = make([]string, 0)
 	for _, link := range boardLinks {
 		if rawBoard, meta, err = client.GetRawDashboardByUID(ctx, link.UID); err != nil {
@@ -140,7 +154,7 @@ func getFolderNameIDMap(client *sdk.Client, ctx context.Context) map[string]int 
 
 //ExportDashboards finds all the dashboards in the configured location and exports them to grafana.
 // if the folde doesn't exist, it'll be created.
-func ExportDashboards(client *sdk.Client, folderFilters []string, query string, conf *viper.Viper) {
+func ExportDashboards(client *sdk.Client, filters DashboardFilter, conf *viper.Viper) {
 	filesInDir := findAllFiles(getResourcePath(conf, "dashboard"))
 	ctx := context.Background()
 	var rawBoard []byte
@@ -150,37 +164,55 @@ func ExportDashboards(client *sdk.Client, folderFilters []string, query string, 
 	var folderId int
 
 	for _, file := range filesInDir {
+		baseFile := filepath.Base(file)
+		baseFile = strings.ReplaceAll(baseFile, ".json", "")
 		if strings.HasSuffix(file, ".json") {
 			if rawBoard, err = ioutil.ReadFile(file); err != nil {
 				log.Println(err)
 				continue
 			}
-			elements := strings.Split(file, "/")
-			if len(elements) >= 2 {
-				folderName = elements[len(elements)-2]
-			}
-			if folderName == "" || folderName == DefaultFolderName {
-				folderId = sdk.DefaultFolderId
-			} else {
-				if val, ok := folderMap[folderName]; ok {
-					folderId = val
-				} else {
-					folder := sdk.Folder{Title: folderName}
-					folder, err = client.CreateFolder(ctx, folder)
-					if err != nil {
-						panic(err)
-					}
-					folderMap[folderName] = folder.ID
-					folderId = folder.ID
-				}
-			}
-
 			var board = make(map[string]interface{})
 			if err = json.Unmarshal(rawBoard, &board); err != nil {
 				log.Println(err)
 				log.Printf("Failed to unmarshall file: %s", file)
 				continue
 			}
+
+			elements := strings.Split(file, "/")
+			if len(elements) >= 2 {
+				folderName = elements[len(elements)-2]
+			}
+			if folderName == "" || folderName == DefaultFolderName {
+				folderId = sdk.DefaultFolderId
+				folderName = DefaultFolderName
+			}
+
+			if folderName == DefaultFolderName {
+				folderId = sdk.DefaultFolderId
+			} else {
+				if val, ok := folderMap[folderName]; ok {
+					folderId = val
+				} else {
+					createFolder := filters.ValidateFolder(folderName)
+					validUid := filters.ValidateDashboard(baseFile)
+
+					if createFolder && validUid {
+						folder := sdk.Folder{Title: folderName}
+						folder, err = client.CreateFolder(ctx, folder)
+						if err != nil {
+							panic(err)
+						}
+						folderMap[folderName] = folder.ID
+						folderId = folder.ID
+					}
+				}
+			}
+
+			//If folder OR slug is filtered, then skip if it doesn't match
+			if !filters.Validate(folderName, baseFile) {
+				continue
+			}
+
 			title, err := jsonpath.Read(board, "$.title")
 
 			rawTitle := fmt.Sprintf("%v", title)
@@ -209,21 +241,25 @@ func ExportDashboards(client *sdk.Client, folderFilters []string, query string, 
 
 //DeleteAllDashboards clears all current dashboards being monitored.  Any folder not white listed
 // will not be affected
-func DeleteAllDashboards(client *sdk.Client, folderFilters []string) []string {
+func DeleteAllDashboards(client *sdk.Client, filter DashboardFilter) []string {
 	ctx := context.Background()
 	var dashboards []string = make([]string, 0)
 
-	filters := DashboardFilter{
-		DashFilter: strings.Join(config.GetDefaultGrafanaConfig().GetMonitoredFolders(), ","),
-	}
-
-	items := ListDashboards(client, &filters, "")
+	items := ListDashboards(client, &filter)
 	for _, item := range items {
-		_, err := client.DeleteDashboardByUID(ctx, item.UID)
-		if err == nil {
-			dashboards = append(dashboards, item.Title)
+		if filter.Validate(item.FolderTitle, item.Slug) {
+			_, err := client.DeleteDashboardByUID(ctx, item.UID)
+			if err == nil {
+				dashboards = append(dashboards, item.Title)
+			}
 		}
 	}
 	return dashboards
 
+}
+
+var quoteRegex *regexp.Regexp
+
+func init() {
+	quoteRegex, _ = regexp.Compile("['\"]+")
 }
